Rename 'attribute' module to 'data' to better conform to Aria Ops conventions
